--- a/src/common/options/global.yaml.in
+++ b/src/common/options/global.yaml.in
@@ -5375,6 +5375,19 @@ options:
   desc: Remove allocation info from RocksDB and store the info in a new allocation file
   default: true
   with_legacy: true
+- name: bluestore_allocation_recovery_threads
+  type: uint
+  level: basic
+  desc: Amount of threads for allocation recovery after OSD crash.
+  long_desc: The optimal value varies. For NVMe DBs may be as large as 8.
+    Value 0 selects legacy recovery.
+    Value 1 denotes new recovery but on single thread.
+  default: 0
+  with_legacy: false
+  flags:
+    - startup
+  see_also:
+    - bluestore_allocation_from_file
 - name: bluestore_debug_inject_allocation_from_file_failure
   type: float
   level: dev