--- a/src/common/options/global.yaml.in
+++ b/src/common/options/global.yaml.in
@@ -5388,6 +5388,19 @@ options:
     - startup
   see_also:
     - bluestore_allocation_from_file
+- name: bluestore_debug_fast_recovery_compare_chance
+  type: float
+  level: dev
+  desc: For testing only. Compare legacy and multithread allocation recovery.
+  long_desc: Chance to perform compare between procedures.
+    Value in range <0..1>; defines chance to perform comparision between legacy
+    and multithreaded allocation recovery. 0 means never compare. 1 means always compare.
+  default: 0
+  with_legacy: false
+  flags:
+    - startup
+  see_also:
+    - bluestore_allocation_recovery_threads
 - name: bluestore_debug_inject_allocation_from_file_failure
   type: float
   level: dev