… src/crimson/os/alienstore/CMakeLists.txt
… src/os/bluestore/CMakeLists.txt
… src/os/bluestore/OnodeScan.cc
… src/os/bluestore/bluestore_tool.cc
… src/os/bluestore/bluestore_types.h
… src/os/bluestore/simple_bitmap.cc
… src/os/bluestore/simple_bitmap.h
… src/test/objectstore/test_bluestore_types.cc
… src/test/objectstore/test_kv.cc